+++ LabelingDevice.cproj
@@ -124,6 +124,9 @@
     </ToolchainSettings>
   </PropertyGroup>
   <ItemGroup>
+    <Compile Include="headers\buttons.h">
+      <SubType>compile</SubType>
+    </Compile>
     <Compile Include="headers\config.h">
       <SubType>compile</SubType>
     </Compile>
@@ -134,6 +137,9 @@
       <SubType>compile</SubType>
     </Compile>
     <Compile Include="main.c">
+      <SubType>compile</SubType>
+    </Compile>
+    <Compile Include="sourceFiles\buttons.c">
       <SubType>compile</SubType>
     </Compile>
     <Compile Include="sourceFiles\lcd_control.c">
